[REFACTOR] useScrap hook 구현
